[TWINFACES-494] feat: refactor fuctional
